neoforge: move launch-plugin injection off Unsafe; drop dead code
The launch-plugin injection touches cpw.mods.modlauncher, which is reachable, so
it now uses ordinary reflection (setAccessible) instead of Unsafe - including the
final-field write of LaunchPluginHandler.plugins, which works via reflection on
the runtime JDK. Unsafe is no longer needed there.

Unsafe cannot be removed entirely: the GAME-layer bridge reaches into
cpw.mods.cl.ModuleClassLoader, whose module (cpw.mods.securejarhandler) does not
`opens cpw.mods.cl` to us, so Field.setAccessible throws InaccessibleObjectException
(probed at runtime). Unsafe field offsets read the object layout, not the value,
so they cross that sealed boundary. ModuleClassLoaderAccess is now scoped to just
that: packageLookup + parentLoaders. Removed its unused generic
readStatic/read/write helpers and the resolvedRoots reader.

Cleanup of earlier non-working attempts: removed the dead
META-INF/services/...ILaunchPluginService file (ModLauncher only discovers launch
plugins from the BOOT layer, so a SERVICE-layer entry is never read - we inject
instead), and corrected AutoModpackCoreMod's stale comment (GameGraphicsBootstrapTrigger
is an idempotent bridge fallback, not a bootstrapper re-fire).

Verified: all five 1.21.1-neoforge autotester scenarios (Sodium, asynclogger,
Connector, Kotlin, CrashAssistant) still PASS.
