+++ loader/neoforge/fml4/src/main/java/pl/skidam/automodpack_loader_core_neoforge/EarlyServiceBridgePlugin.java
@@ -5,6 +5,7 @@ import cpw.mods.modlauncher.api.NamedPath;
 import cpw.mods.modlauncher.serviceapi.ILaunchPluginService;
 import org.objectweb.asm.Type;
 
+import java.lang.reflect.Field;
 import java.util.EnumSet;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -27,8 +28,9 @@ import static pl.skidam.automodpack_core.Constants.LOGGER;
  * single, already-bootstrapped child-layer copy. No GAME-library copy, no split static state.
  *
  * <p>ModLauncher only discovers launch plugins from the BOOT module layer, while AutoModpack ships on
- * the SERVICE layer, so our {@code META-INF/services} entry is never picked up on its own. We inject
- * this plugin into ModLauncher's launch-plugin map by reflection instead - see {@link #ensureRunsFirst()}.
+ * the SERVICE layer, so a {@code META-INF/services} entry would never be picked up - we don't ship one.
+ * Instead we inject this plugin into ModLauncher's launch-plugin map by reflection - see
+ * {@link #ensureRunsFirst()}.
  */
 @SuppressWarnings("unused")
 public class EarlyServiceBridgePlugin implements ILaunchPluginService {
@@ -40,9 +42,9 @@ public class EarlyServiceBridgePlugin implements ILaunchPluginService {
      *
      * <p>ModLauncher discovers {@link ILaunchPluginService}s only from the BOOT module layer
      * ({@code LaunchPluginHandler}'s constructor does {@code ServiceLoader.load(BOOT, ...)}).
-     * AutoModpack ships on the SERVICE layer, so our {@code META-INF/services} entry is never seen -
-     * confirmed empirically: the launch-plugin map only ever holds NeoForge's built-in BOOT plugins
-     * (mixin, slf4jfixer, ...) and never ours. So we inject our own instance into that map by
+     * AutoModpack ships on the SERVICE layer, so a {@code META-INF/services} entry would never be
+     * seen - confirmed empirically: the launch-plugin map only ever holds NeoForge's built-in BOOT
+     * plugins (mixin, slf4jfixer, ...) and never ours. So we inject our own instance into that map by
      * reflection, from our {@code GraphicsBootstrapper} (the early-window phase, well before the
      * launch). {@code announceLaunch} then calls our {@link #initializeLaunch} - which runs the
      * GAME-layer bridge - after the GAME {@code TransformingClassLoader} is built but before the game
@@ -54,20 +56,46 @@ public class EarlyServiceBridgePlugin implements ILaunchPluginService {
      */
     static void ensureRunsFirst() {
         try {
-            Object launcher = ModuleClassLoaderAccess.readStatic(Class.forName("cpw.mods.modlauncher.Launcher"), "INSTANCE");
+            // Plain reflection is enough here: unlike cpw.mods.securejarhandler (which seals
+            // cpw.mods.cl, forcing the GAME-layer bridge onto Unsafe), the cpw.mods.modlauncher
+            // module is reachable, so we can read/write its fields with setAccessible.
+            Object launcher = Class.forName("cpw.mods.modlauncher.Launcher").getField("INSTANCE").get(null);
             if (launcher == null) return;
-            Object pluginHandler = ModuleClassLoaderAccess.read(launcher, "launchPlugins");
+            Object pluginHandler = readField(launcher, "launchPlugins");
             if (pluginHandler == null) return;
             @SuppressWarnings("unchecked")
-            Map<String, Object> plugins = (Map<String, Object>) ModuleClassLoaderAccess.read(pluginHandler, "plugins");
+            Map<String, Object> plugins = (Map<String, Object>) readField(pluginHandler, "plugins");
             if (plugins == null || plugins instanceof PriorityFirstPluginMap) return;
             PriorityFirstPluginMap ordered = new PriorityFirstPluginMap(NAME, plugins);
             ordered.putIfAbsent(NAME, new EarlyServiceBridgePlugin());
-            ModuleClassLoaderAccess.write(pluginHandler, "plugins", ordered);
-            LOGGER.info("[AutoModpack] Injected the early-service bridge as ModLauncher launch plugin '{}' and ordered it first (ModLauncher discovers launch plugins only from the BOOT layer, so our SERVICE-layer service file is otherwise ignored)", NAME);
+            writeField(pluginHandler, "plugins", ordered);
+            LOGGER.info("[AutoModpack] Injected the early-service bridge as ModLauncher launch plugin '{}' and ordered it first (ModLauncher discovers launch plugins only from the BOOT layer, so a SERVICE-layer service file would never be seen)", NAME);
         } catch (Throwable t) {
             LOGGER.warn("[AutoModpack] Could not inject the early-service bridge launch plugin; in-place graphics-service mods may fail", t);
         }
+    }
+
+    private static Object readField(Object owner, String name) throws Exception {
+        Field field = findField(owner.getClass(), name);
+        field.setAccessible(true);
+        return field.get(owner);
+    }
+
+    private static void writeField(Object owner, String name, Object value) throws Exception {
+        Field field = findField(owner.getClass(), name);
+        field.setAccessible(true);
+        field.set(owner, value);
+    }
+
+    private static Field findField(Class<?> from, String name) throws NoSuchFieldException {
+        for (Class<?> c = from; c != null; c = c.getSuperclass()) {
+            try {
+                return c.getDeclaredField(name);
+            } catch (NoSuchFieldException ignored) {
+                // try superclass
+            }
+        }
+        throw new NoSuchFieldException(name);
     }
 
     /**
