--- loader/neoforge/fml4/src/main/resources/META-INF/services/cpw.mods.modlauncher.serviceapi.ILaunchPluginService
@@ -1,1 +1,0 @@
-pl.skidam.automodpack_loader_core.EarlyServiceBridgePlugin
